pyproject.toml
[tool.poetry]
name = "hse-hw3-ap-url-shortener"
version = "0.1.0"
description = ""
authors = ["TheSeems <me@theseems.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
sqlmodel = "^0.0.24"
uvicorn = "^0.34.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.10"
redis = ">=4.5.5"
apscheduler = ">=3.10.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
types-passlib = "^1.7.7.20250326"
pylint = "^3.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
